requirements.txt
fastapi==0.115.0
uvicorn[standard]==0.30.6
numpy==1.26.4
opencv-python-headless==4.9.0.80
faiss-cpu==1.8.0.post1
insightface==0.7.3
onnxruntime==1.18.1
torch==2.3.1
torchvision==0.18.1
timm==1.0.8
scikit-learn==1.5.1
python-multipart==0.0.9
pyjwt==2.9.0
httpx==0.27.0
